games: Give access to joypads
GNOME Games currently uses udev for joypad enumerations, which doesn't
work inside a flatpak.

It has a fallback code which searches inside /dev. This is pretty
terrible as it forces to poke a giant hole in the sandbox, giving Games
access to all devices.

However the alternative is not being able to play games.

Eventually, we'll need something like a joypad portal, and we'll need to
port Games to it, but in the meantime, this is the best we can do.